refactor fileUtil.ts to a load module
