+++ app/src/lib/load/index.ts
@@ -1,0 +1,3 @@
+export { default as Parquet } from "./parquet";
+export { default as CSV } from "./csv";
+export { default as Excel } from "./excel";
